requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-heroku==0.3.1
djangorestframework==3.11.0
et-xmlfile==1.0.1
googlemaps==4.1.0
gunicorn==20.0.4
idna==2.8
jdcal==1.4.1
matplotlib==3.1.2
numpy==1.18.1
openpyxl==3.0.3
pandas==0.25.3
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
smsapi-client==2.4.0
sqlparse==0.3.0
urllib3==1.25.8
WeasyPrint==51
xlrd==1.2.0
